moving test_helpers to testing/helpers.py (it contains no tests, but helping functions)
