+++ nibabel/tests/test_brikhead.py
@@ -16,10 +16,9 @@ from .. import brikhead
 
 import pytest
 from numpy.testing import assert_array_equal
-from ..testing_pytest import data_path
+from ..testing_pytest import data_path, assert_data_similar
 
 from .test_fileslice import slicer_samples
-from .test_helpers import assert_data_similar
 
 EXAMPLE_IMAGES = [
     dict(
